backend/requirements.txt
# alembic            1.14.0
# annotated-types    0.7.0
# anyio              4.7.0
# asn1crypto         1.5.1
# bcrypt             4.2.1
# cachetools         5.5.0
# certifi            2024.12.14
# charset-normalizer 3.4.1
# click              8.1.8
# colorama           0.4.6
# fastapi            0.115.6
# google-auth        2.37.0
# h11                0.14.0
# idna               3.10
# Mako               1.3.8
# MarkupSafe         3.0.2
# passlib            1.7.4
# pg8000             1.31.2
# pip                24.3.1
# pyasn1             0.6.1
# pyasn1_modules     0.4.1
# pydantic           2.10.4
# pydantic_core      2.27.2
# PyJWT              2.10.1
# python-dateutil    2.9.0.post0
# python-multipart   0.0.20
# requests           2.32.3
# rsa                4.9
# scramp             1.4.5
# six                1.17.0
# sniffio            1.3.1
# SQLAlchemy         2.0.36
# starlette          0.41.3
# typing             3.7.4.3
# typing_extensions  4.12.2
# urllib3            2.3.0
# uvicorn            0.34.0


alembic            1.14.0
annotated-types    0.7.0
anyio              4.8.0
asn1crypto         1.5.1
cachetools         5.5.0
certifi            2024.12.14
charset-normalizer 3.4.1
click              8.1.8
colorama           0.4.6
fastapi            0.115.6
google-auth        2.37.0
h11                0.14.0
idna               3.10
Mako               1.3.8
MarkupSafe         3.0.2
passlib            1.7.4
pg8000             1.31.2
pip                24.3.1
pyasn1             0.6.1
pyasn1_modules     0.4.1
pydantic           2.10.4
pydantic_core      2.27.2
PyJWT              2.10.1
python-dateutil    2.9.0.post0
python-multipart   0.0.20
requests           2.32.3
rsa                4.9
scramp             1.4.5
six                1.17.0
sniffio            1.3.1
SQLAlchemy         2.0.36
starlette          0.41.3
typing             3.7.4.3
typing_extensions  4.12.2
urllib3            2.3.0
uvicorn            0.34.0
